chore(http): split types into modules
- Request
- Response
- Scope
