+++ src/Http/Request/types.ts
@@ -1,0 +1,11 @@
+import type { File } from 'formidable';
+import type { Request } from 'koa';
+
+export type UploadedFile = File;
+export type UploadedFilesMap = Record<string, UploadedFile | UploadedFile[]>;
+
+export interface HttpRequest extends Request {
+  body?: Record<string, unknown>;
+  files?: UploadedFilesMap;
+  params: Record<string, unknown>;
+}
