+++ src/Http/Scope/types.ts
@@ -1,0 +1,11 @@
+import type { Context, DefaultState, Next } from 'koa';
+import { type HttpRequest, type HttpResponse } from '@/Http';
+
+export interface HttpScope extends Context {
+  request: HttpRequest;
+  response: HttpResponse;
+}
+
+export type NextMiddleware = Next;
+export type HttpMiddleware = (scope: HttpScope, next: NextMiddleware) => Promise<unknown>;
+export type AppState = DefaultState;
