+++ src/Http/index.ts
@@ -1,1 +1,3 @@
-export type * from './types';
+export type * from './Request';
+export type * from './Response';
+export type * from './Scope';
